chore: move logic back into dedicated modules
